% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arReceivable.R
\name{arReceivable_ERPSync}
\alias{arReceivable_ERPSync}
\title{应收单同步}
\usage{
arReceivable_ERPSync()
}
\arguments{
\item{dms_token}{}

\item{FBillNo}{}
}
\value{

}
\description{
应收单同步
}
\examples{
arReceivable_ERPSync()
}
